fix: context hang issue
