+++ example/crawler_report.json
@@ -1,14 +1,14 @@
 {
-  "seed": "https://example.com",
+  "seed": "https://example.com/",
   "max_depth": 3,
   "blacklist": [""],
   "network_info": {
     "example.com": [
       {
-        "paths": [""],
+        "paths": ["/"],
         "remote_addr": ["93.184.216.34:443"],
         "dns_addrs": ["2606:2800:220:1:248:1893:25c8:1946", "93.184.216.34"],
-        "avg_response_ms": 720
+        "avg_response_ms": 729
       }
     ],
     "www.iana.org": [
@@ -16,7 +16,7 @@
         "paths": ["/domains/example"],
         "remote_addr": ["192.0.33.8:80"],
         "dns_addrs": ["192.0.33.8", "2620:0:2d0:200::b:8"],
-        "avg_response_ms": 1240
+        "avg_response_ms": 1187
       }
     ],
     "www.icann.org": [
@@ -24,7 +24,7 @@
         "paths": ["/", "/privacy/policy", "/privacy/tos"],
         "remote_addr": ["192.0.32.7:443"],
         "dns_addrs": ["192.0.32.7", "2620:0:2d0:200::7"],
-        "avg_response_ms": 1243
+        "avg_response_ms": 1301
       }
     ]
   },
@@ -51,19 +51,19 @@
         "https://www.icann.org/en/announcements/details/icann-publishes-fy23-annual-report-and-audited-financial-statements-04-10-2023-en",
         "https://www.icann.org/en/announcements/details/nextgenicann-applications-for-icannn79-extended-to-early-november-2023-16-10-2023-en",
         "https://www.icann.org/en/blogs",
+        "https://www.icann.org/en/blogs/details/chairs-blog-preview-of-the-october-board-workshop-and-icann78-17-10-2023-en",
         "https://www.icann.org/en/public-comment/proceeding/pilot-holistic-review-revised-terms-of-reference-tor-28-09-2023",
         "https://www.icann.org/en/public-comment/proceeding/proposed-updates-to-existing-rights-protection-mechanisms-documentation-24-08-2023",
         "https://www.icann.org/public-comments",
         "https://www.icann.org/resources/pages/octo-publications-2019-05-24-en ",
         "https://www.icann.org/resources/pages/social-media-2020-12-14-en",
-        "https://www.icann.org/sites/default/files/meetings/icann78-policy-outlook-report-06oct23-en.pdf",
         "https://www.instagram.com/icannorg/",
         "https://www.linkedin.com/company/icann",
         "https://www.soundcloud.com/icann/",
         "https://www.youtube.com/user/ICANNnews"
       ]
     },
-    "https://example.com": {
+    "https://example.com/": {
       "depth": 0,
       "links": ["https://www.iana.org/domains/example"]
     },
