+++ internal/crawler/option.go
@@ -1,24 +1,10 @@
 package crawler
 
 import (
-	"github.com/yusufaine/cs3203-g46-crawler/pkg/rhttp"
 	"golang.org/x/time/rate"
 )
 
 type CrawlerOption func(*Crawler)
-
-// Hosts specified in the blacklist will not be crawled
-func WithBlacklist(blacklist map[string]struct{}) CrawlerOption {
-	return func(c *Crawler) {
-		c.HostBlacklist = blacklist
-	}
-}
-
-func WithRHttpClient(hc *rhttp.Client) CrawlerOption {
-	return func(c *Crawler) {
-		c.hc = hc
-	}
-}
 
 // Determines how links are extracted from the response body.
 // See: linkextractor.go
